requirements/main.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --output-file requirements/main.txt requirements/main.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
click==8.1.7
    # via
    #   safir
    #   uvicorn
cryptography==43.0.1
    # via
    #   pyjwt
    #   safir
fastapi==0.114.0
    # via
    #   -r requirements/main.in
    #   safir
gidgethub==5.3.0
    # via safir
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via safir
httpx-sse==0.4.0
    # via -r requirements/main.in
idna==3.8
    # via
    #   anyio
    #   httpx
pycparser==2.22
    # via cffi
pydantic==2.9.0
    # via
    #   -r requirements/main.in
    #   fastapi
    #   pydantic-settings
    #   safir
pydantic-core==2.23.2
    # via
    #   pydantic
    #   safir
pydantic-settings==2.4.0
    # via -r requirements/main.in
pyjwt==2.9.0
    # via gidgethub
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via
    #   -r requirements/main.in
    #   uvicorn
rubin-nublado-client @ git+https://github.com/lsst-sqre/nublado@0ab1f77f60489d1df22bfb04ec64ee703ba594f4#subdirectory=client
    # via -r requirements/main.in
safir==6.3.0
    # via -r requirements/main.in
safir-logging==6.3.0
    # via safir
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.38.4
    # via
    #   -r requirements/main.in
    #   fastapi
    #   safir
structlog==24.4.0
    # via
    #   safir
    #   safir-logging
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via pydantic
uritemplate==4.1.1
    # via gidgethub
uvicorn==0.30.6
    # via -r requirements/main.in
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.0.1
    # via uvicorn
